Simplifications for clarity & bump to Jetty 9.
Change jetty8 dependency to jetty9 in pom.xml and update ApplicationModule to reflect this.
Change HttpRoutes injection to annotation on private member instead of constructor + remove constructor to improve readability of code as an example.
Remove the string formatting of the route as a hardcoded string will suffice for this example.
Rename PingEndpoint -> PingProcessor to reflect that it's a Camel processor.
Remove intermediate variable in PingProcessor because it doesn't add anything to readability here.
Hard code values in HttpRoutesTest, again to make this more readable as example code.